refactor: move io-error class
IOError is closely associated with file-io. Move it there.
