+++ src/common-errors.ts
@@ -2,21 +2,6 @@ import { CustomError } from "ts-custom-error";
 import { EditorVersion } from "./domain/editor-version";
 
 /**
- * Generic IO error for when interacting with the file-system failed.
- */
-export class IOError extends CustomError {
-  private readonly _class = "CustomError";
-  constructor(
-    /**
-     * The actual error that caused the failure.
-     */
-    cause?: NodeJS.ErrnoException
-  ) {
-    super("An interaction with the file-system caused an error.", { cause });
-  }
-}
-
-/**
  * Error for when the packument was not found.
  */
 export class PackumentNotFoundError extends CustomError {
