+++ test/env.test.ts
@@ -6,8 +6,8 @@ import { tryLoadProjectVersion } from "../src/io/project-version-io";
 import { Err, Ok } from "ts-results-es";
 import { tryGetUpmConfigDir } from "../src/io/upm-config-io";
 import fs from "fs";
-import { NotFoundError } from "../src/io/file-io";
-import { FileParseError, IOError } from "../src/common-errors";
+import { IOError, NotFoundError } from "../src/io/file-io";
+import { FileParseError } from "../src/common-errors";
 import { makeEditorVersion } from "../src/domain/editor-version";
 import { NoWslError } from "../src/io/wls";
 import { mockUpmConfig } from "./upm-config-io.mock";
