+++ test/npmrc-token-update-service.test.ts
@@ -6,9 +6,9 @@ import {
 import { AsyncResult, Err, Ok } from "ts-results-es";
 import { RequiredEnvMissingError } from "../src/io/upm-config-io";
 import { tryUpdateUserNpmrcToken } from "../src/services/npmrc-token-update";
-import { IOError } from "../src/common-errors";
 import { emptyNpmrc, setToken } from "../src/domain/npmrc";
 import { exampleRegistryUrl } from "./data-registry";
+import { IOError } from "../src/io/file-io";
 
 const exampleNpmrcPath = "/users/someuser/.npmrc";
